feat: change i18n structure, set gctime 0 to review get
