+++ app/i18n/index.ts
@@ -1,12 +1,23 @@
-import { type I18nBase as _I18nBase, i18nBaseNamespaces } from "./resources/_base"
-import { i18nEn } from "./resources/en"
-import { i18nKo } from "./resources/ko"
+import { i18nEn as i18nEnBundle } from "./locales/en"
+import { i18nKo as i18nKoBundle } from "./locales/ko"
 
-export const namespaces = i18nBaseNamespaces
-
-export const resources = {
-  ko: i18nKo,
-  en: i18nEn,
+export const i18nBase = {
+    common: i18nKoBundle,
 }
 
-export type I18nBase = _I18nBase
+export type I18nBase = typeof i18nBase
+
+export const i18nKo: I18nBase = {
+    common: i18nKoBundle,
+}
+
+export const i18nEn: I18nBase = {
+    common: i18nEnBundle,
+}
+
+export const namespaces = Object.keys(i18nBase)
+
+export const resources = {
+    ko: i18nKo,
+    en: i18nEn,
+}
