+++ app/i18n/locales/en/common.ts
@@ -1,0 +1,130 @@
+export const common = {
+    class: "Class",
+    professor: "Professor",
+    description: "Description",
+    grade: "Grade",
+    load: "Load",
+    speech: "Speech",
+    upload: "Upload",
+    numClasses: "Lec. Hours",
+    numLabs: "Lab Hours",
+    credit: "Credits",
+    cancel: "Cancel",
+    save: "Save",
+    edit: "Edit",
+    hours: "Hours",
+    minutes: "Min",
+    seconds: "Sec",
+    am: "AM",
+    pm: "PM",
+    year: "year",
+    language: "Language",
+    type: {
+        basic: "Basic",
+        major: "Major",
+        doubleMajor: "Double Major",
+        minor: "Minor",
+        advancedMajor: "Advanced Major",
+        interdisciplinaryMajor: "Interdisciplinary Major",
+        research: "Research",
+        general: "General",
+        basicRequired: "Basic Required",
+        basicElective: "Basic Elective",
+        majorRequired: "Major Required",
+        majorElective: "Major Elective",
+        generalRequired: "General Required",
+        humanitiesSocialElective: "Humanities & Social Elective",
+        humanities: "Humanities",
+        thesisStudy: "Thesis Study",
+        individualStudy: "Individual Study",
+        otherElective: "Other Elective",
+        unclassified: "Unclassified",
+        etc: "Others",
+        basicRequiredShort: "BR",
+        basicElectiveShort: "BE",
+        majorRequiredShort: "MR",
+        majorElectiveShort: "ME",
+        humanitiesSocialElectiveShort: "HSE",
+        mandatoryGeneralCourseShort: "MGC",
+        otherElectiveShort: "OE",
+        generalRequiredShort: "GR",
+        electiveGraduateShort: "EG",
+        etcShort: "ETC",
+        totalCredit: "Credit",
+        totalAu: "AU",
+    },
+    department: {
+        hssShort: "HSS",
+        aeShort: "AE",
+        bisShort: "BiS",
+        bsShort: "BS",
+        cbeShort: "CBE",
+        ceShort: "CE",
+        chShort: "CH",
+        csShort: "CS",
+        eeShort: "EE",
+        idShort: "ID",
+        ieShort: "IE",
+        masShort: "MAS",
+        meShort: "ME",
+        nqeShort: "NQE",
+        btmShort: "BTM",
+        msShort: "MS",
+        phShort: "PH",
+        tsShort: "TS",
+        ssShort: "SS",
+        bcsShort: "BCS",
+        etcShort: "ETC",
+    },
+    level: {
+        "100sShort": "100s",
+        "200sShort": "200s",
+        "300sShort": "300s",
+        "400sShort": "400s",
+        "500sShort": "500s",
+        "600sShort": "600s",
+        "700sShort": "700s",
+        "800sShort": "800s",
+        "900sShort": "900s",
+        etcShort: "ETC",
+    },
+    term: {
+        "3yearsShort": "In-3-Years",
+        "1yearShort": "In-1-Year",
+        thisSemesterShort: "This-Semester",
+    },
+    professors: {
+        over: " and ",
+        people: " others",
+    },
+    review: {
+        writingPlaceholder: "Please describe your evaluation of grades, load, etc.",
+        like: "Like ",
+    },
+    search: {
+        search: "Search",
+        cancel: "Cancel",
+        level: "Grade",
+        department: "Department",
+        term: "Term",
+        all: "All",
+        type: "Groups",
+        placeholder: "Enter keywords",
+        empty: "Please select at least one search filter",
+    },
+    semesters: {
+        spring: "Spring",
+        summer: "Summer",
+        fall: "Fall",
+        winter: "Winter",
+    },
+    days: {
+        monday: "Mon",
+        tuesday: "Tue",
+        wednesday: "Wed",
+        thursday: "Thu",
+        friday: "Fri",
+        saturday: "Sat",
+        sunday: "Sun",
+    },
+}
