+++ app/i18n/locales/en/credits.ts
@@ -1,0 +1,6 @@
+export const credits = {
+    credits: "Credits",
+    license: "License",
+    privacyPolicy: "Privacy Policy",
+    contact: "Contact Us",
+}
