+++ app/i18n/locales/en/dictionary.ts
@@ -1,0 +1,19 @@
+export const dictionary = {
+    courseHistory: "Course History",
+    notOffered: "Not offered",
+    courseCountInfo:
+        "Total<space/><bold>{{count}}</bold><space/>courses<space/>(<icon/>: offered this semester)",
+    sort: "Sort",
+    noResults: "No results",
+    sortOptions: {
+        code: "By Course Code",
+        popularity: "By Popularity",
+        studentCount: "By Enrollment",
+    },
+    review: "Course Reviews",
+    reviewLanguage: "Language",
+    reviewLanguageOptions: {
+        all: "All",
+        english: "English",
+    },
+}
