+++ app/i18n/locales/en/header.ts
@@ -1,0 +1,5 @@
+export const header = {
+    dictionary: "Dictionary",
+    writeReviews: "Write Reviews",
+    timetable: "Timetable",
+}
