+++ app/i18n/locales/en/index.ts
@@ -1,0 +1,17 @@
+import { account } from "./account"
+import { common } from "./common"
+import { credits } from "./credits"
+import { dictionary } from "./dictionary"
+import { header } from "./header"
+import { main } from "./main"
+import { writeReviews } from "./writeReviews"
+
+export const i18nEn = {
+    account,
+    common,
+    credits,
+    dictionary,
+    header,
+    main,
+    writeReviews,
+}
