+++ app/i18n/locales/en/writeReviews.ts
@@ -1,0 +1,24 @@
+export const writeReviews = {
+    mySummary: {
+        title: "Taken Lectures",
+        written: "My Reviews",
+        likes: "likes",
+    },
+    write: {
+        title: "Write Reviews - {{lectureName}}",
+        related: "Related Reviews - {{lectureName}}",
+    },
+    hallOfFameFeed: {
+        title: "Hall of Fame - {{year}}<space/>{{semester}}",
+        total: "Total Reviews",
+    },
+    likedReviews: {
+        title: "My Liked Reviews",
+    },
+    tabs: {
+        write: "Write Reviews",
+        hallOfFameFeed: "Hall of Fame",
+        recentFeed: "Fresh Reviews",
+        liked: "My Liked Reviews",
+    },
+}
