+++ app/i18n/locales/ko/header.ts
@@ -1,0 +1,5 @@
+export const header = {
+    dictionary: "과목사전",
+    timetable: "모의시간표",
+    writeReviews: "과목후기 작성하기",
+}
