+++ app/i18n/locales/ko/writeReviews.ts
@@ -1,0 +1,24 @@
+export const writeReviews = {
+    mySummary: {
+        title: "내가 들은 과목",
+        written: "작성 후기",
+        likes: "추천",
+    },
+    write: {
+        title: "후기 작성 - {{lectureName}}",
+        related: "연관 후기 - {{lectureName}}",
+    },
+    hallOfFameFeed: {
+        title: "명예의 전당 - {{year}}<space/>{{semester}}",
+        total: "전체 후기",
+    },
+    likedReviews: {
+        title: "좋아요한 후기",
+    },
+    tabs: {
+        write: "후기 작성하기",
+        recentFeed: "따끈따끈 후기",
+        hallOfFameFeed: "명예의 전당",
+        liked: "좋아요한 후기",
+    },
+}
