--- app/i18n/resources/en.ts
@@ -1,6 +1,0 @@
-import { commonI18nEn } from "../variables/en"
-import { type I18nBase } from "./_base"
-
-export const i18nEn: I18nBase = {
-  common: commonI18nEn,
-}
