--- app/i18n/resources/ko.ts
@@ -1,6 +1,0 @@
-import { commonI18nKo } from "../variables/ko"
-import { type I18nBase } from "./_base"
-
-export const i18nKo: I18nBase = {
-    common: commonI18nKo,
-}
