--- app/i18n/variables/_base.ts
@@ -1,234 +1,0 @@
-export const commonI18nBase = {
-    credits: {
-        credits: "만든 사람들",
-        license: "이용 약관",
-        privacyPolicy: "개인정보취급방침",
-        contact: "문의하기",
-    },
-    common: {
-        class: "분류",
-        professor: "교수",
-        description: "설명",
-        grade: "성적",
-        load: "널널",
-        speech: "강의",
-        upload: "업로드",
-        numClasses: "강의시간",
-        numLabs: "실험",
-        credit: "학점",
-        cancel: "취소",
-        save: "저장",
-        edit: "수정",
-        hours: "시간",
-        minutes: "분",
-        seconds: "초",
-        am: "오전",
-        pm: "오후",
-        year: "연도",
-        language: "언어",
-        type: {
-            basic: "기초",
-            major: "전공",
-            doubleMajor: "복수전공",
-            minor: "부전공",
-            advancedMajor: "심화전공",
-            interdisciplinaryMajor: "자유융합전공",
-            research: "연구",
-            general: "교양",
-            basicRequired: "기초필수",
-            basicElective: "기초선택",
-            majorRequired: "전공필수",
-            majorElective: "전공선택",
-            generalRequired: "교양필수",
-            humanitiesSocialElective: "인문사회선택",
-            humanities: "인문사회선택",
-            thesisStudy: "졸업연구",
-            individualStudy: "개별연구",
-            otherElective: "자유선택",
-            unclassified: "미분류",
-            etc: "기타",
-            basicRequiredShort: "기필",
-            basicElectiveShort: "기선",
-            majorRequiredShort: "전필",
-            majorElectiveShort: "전선",
-            humanitiesSocialElectiveShort: "인선",
-            mandatoryGeneralCourseShort: "교필",
-            otherElectiveShort: "자선",
-            generalRequiredShort: "공통",
-            electiveGraduateShort: "석박",
-            etcShort: "기타",
-            totalCredit: "학점",
-            totalAu: "AU",
-        },
-        department: {
-            hssShort: "인문",
-            aeShort: "항공",
-            bisShort: "바공",
-            bsShort: "생명",
-            cbeShort: "생화공",
-            ceShort: "건환",
-            chShort: "화학",
-            csShort: "전산",
-            eeShort: "전자",
-            idShort: "산디",
-            ieShort: "산공",
-            masShort: "수리",
-            meShort: "기계",
-            nqeShort: "원양",
-            btmShort: "기경",
-            msShort: "신소재",
-            phShort: "물리",
-            tsShort: "융인",
-            ssShort: "반시공",
-            bcsShort: "뇌인지",
-            etcShort: "기타",
-        },
-        level: {
-            "100sShort": "100번대",
-            "200sShort": "200번대",
-            "300sShort": "300번대",
-            "400sShort": "400번대",
-            "500sShort": "500번대",
-            "600sShort": "600번대",
-            "700sShort": "700번대",
-            "800sShort": "800번대",
-            "900sShort": "900번대",
-            etcShort: "기타",
-        },
-        term: {
-            "3yearsShort": "3년이내",
-            "1yearShort": "1년이내",
-            thisSemesterShort: "이번학기",
-        },
-        professors: {
-            over: " 외 ",
-            people: "명",
-        },
-        review: {
-            writingPlaceholder: "학점, 로드 등의 평가에 대하여 설명해주세요.",
-            like: "좋아요",
-        },
-        search: {
-            search: "검색",
-            cancel: "취소",
-            level: "학년",
-            department: "학과",
-            term: "기간",
-            all: "전체",
-            type: "분류",
-            placeholder: "과목명, 교수명을 등을 검색해보세요",
-            empty: "검색 조건을 하나 이상 선택해주세요",
-        },
-        semesters: {
-            spring: "봄",
-            summer: "여름",
-            fall: "가을",
-            winter: "겨울",
-        },
-        days: {
-            monday: "월요일",
-            tuesday: "화요일",
-            wednesday: "수요일",
-            thursday: "목요일",
-            friday: "금요일",
-            saturday: "토요일",
-            sunday: "일요일",
-        },
-    },
-    header: {
-        dictionary: "과목사전",
-        timetable: "모의시간표",
-        writeReviews: "과목후기 작성하기",
-    },
-    main: {
-        hisTimeTable: "<name>{{name}}</name><normal>님의 시간표</normal>",
-        schedule: {
-            title: "{{content}}까지",
-            link: "학사 시스템 바로가기",
-            contents: {
-                beginning: "개강",
-                end: "종강",
-                courseRegistrationPeriodStart: "수강신청기간 시작",
-                courseRegistrationPeriodEnd: "수강신청기간 종료",
-                courseAddDropPeriodEnd: "수강변경기간 종료",
-                courseDropDeadline: "수강취소 마감",
-                courseEvaluationDeadline: "강의평가 마감",
-                gradePosting: "성적게시",
-            },
-        },
-        reviewSection: {
-            title: "{{lectureName}} 강의는 어땠나요?",
-            placeholder: "학점, 로드 등의 평가에 대해 설명해주세요.",
-            writeAnother: "다른 과목 후기 쓰기",
-        },
-        recentFeed: {
-            title: "<bold>따끈따끈</bold><space/><normal>과목후기</normal>",
-        },
-        likedMajorFeed: {
-            title: "<bold>사랑받는</bold><space/><normal>과목후기</normal>",
-        },
-        hallOfFameFeed: {
-            title: "<bold>명예의</bold><space/><normal>전당</normal>",
-        },
-        scheduleFeed: {
-            title: "주요 일정",
-        },
-    },
-    dictionary: {
-        courseHistory: "개설이력",
-        notOffered: "미개설",
-        courseCountInfo: "총 <bold>{{count}}</bold>개 과목 (<icon/>: 이번 학기 개설)",
-        sort: "정렬",
-        noResults: "결과 없음",
-        sortOptions: {
-            code: "과목코드순",
-            popularity: "인기순",
-            studentCount: "수강자 많은 순",
-        },
-        review: "과목 후기",
-        reviewLanguage: "언어",
-        reviewLanguageOptions: {
-            all: "전체",
-            english: "영어",
-        },
-    },
-    account: {
-        title: "내 정보",
-        name: "이름",
-        email: "메일",
-        schoolId: "학번",
-        course: "과정",
-        major: "전공",
-        interestedDepartments: "관심 학과",
-        logout: "로그아웃",
-        sparcsssoCaption: "이 정보는 <bold>SPARCS SSO</bold>에서 변경할 수 있습니다",
-        kaistPortalCaption:
-            "이 정보는 KAIST Portal과 연동됩니다. 정보가 일치하지 않을 경우 <bold>otlplus@sparcs.org</bold>로 문의해주세요",
-    },
-    writeReviews: {
-        mySummary: {
-            title: "내가 들은 과목",
-            written: "작성 후기",
-            likes: "추천",
-        },
-        write: {
-            title: "후기 작성 - {{lectureName}}",
-            related: "연관 후기 - {{lectureName}}",
-        },
-        hallOfFameFeed: {
-            title: "명예의 전당 - {{year}}<space/>{{semester}}",
-            total: "전체 후기",
-        },
-        likedReviews: {
-            title: "좋아요한 후기",
-        },
-        tabs: {
-            write: "후기 작성하기",
-            recentFeed: "따끈따끈 후기",
-            hallOfFameFeed: "명예의 전당",
-            liked: "좋아요한 후기",
-        },
-    },
-}
-
-export type CommonI18nBaseType = typeof commonI18nBase
