--- app/i18n/variables/en.ts
@@ -1,239 +1,0 @@
-import { type CommonI18nBaseType, commonI18nBase } from "./_base"
-
-export const commonI18nEn: CommonI18nBaseType = {
-    ...commonI18nBase,
-    credits: {
-        credits: "Credits",
-        license: "License",
-        privacyPolicy: "Privacy Policy",
-        contact: "Contact Us",
-    },
-    common: {
-        class: "Class",
-        professor: "Professor",
-        description: "Description",
-        grade: "Grade",
-        load: "Load",
-        speech: "Speech",
-        upload: "Upload",
-        numClasses: "Lec. Hours",
-        numLabs: "Lab Hours",
-        credit: "Credits",
-        cancel: "Cancel",
-        save: "Save",
-        edit: "Edit",
-        hours: "Hours",
-        minutes: "Min",
-        seconds: "Sec",
-        am: "AM",
-        pm: "PM",
-        year: "year",
-        language: "Language",
-        type: {
-            basic: "Basic",
-            major: "Major",
-            doubleMajor: "Double Major",
-            minor: "Minor",
-            advancedMajor: "Advanced Major",
-            interdisciplinaryMajor: "Interdisciplinary Major",
-            research: "Research",
-            general: "General",
-            basicRequired: "Basic Required",
-            basicElective: "Basic Elective",
-            majorRequired: "Major Required",
-            majorElective: "Major Elective",
-            generalRequired: "General Required",
-            humanitiesSocialElective: "Humanities & Social Elective",
-            humanities: "Humanities",
-            thesisStudy: "Thesis Study",
-            individualStudy: "Individual Study",
-            otherElective: "Other Elective",
-            unclassified: "Unclassified",
-            etc: "Others",
-            basicRequiredShort: "BR",
-            basicElectiveShort: "BE",
-            majorRequiredShort: "MR",
-            majorElectiveShort: "ME",
-            humanitiesSocialElectiveShort: "HSE",
-            mandatoryGeneralCourseShort: "MGC",
-            otherElectiveShort: "OE",
-            generalRequiredShort: "GR",
-            electiveGraduateShort: "EG",
-            etcShort: "ETC",
-            totalCredit: "Credit",
-            totalAu: "AU",
-        },
-        department: {
-            hssShort: "HSS",
-            aeShort: "AE",
-            bisShort: "BiS",
-            bsShort: "BS",
-            cbeShort: "CBE",
-            ceShort: "CE",
-            chShort: "CH",
-            csShort: "CS",
-            eeShort: "EE",
-            idShort: "ID",
-            ieShort: "IE",
-            masShort: "MAS",
-            meShort: "ME",
-            nqeShort: "NQE",
-            btmShort: "BTM",
-            msShort: "MS",
-            phShort: "PH",
-            tsShort: "TS",
-            ssShort: "SS",
-            bcsShort: "BCS",
-            etcShort: "ETC",
-        },
-        level: {
-            "100sShort": "100s",
-            "200sShort": "200s",
-            "300sShort": "300s",
-            "400sShort": "400s",
-            "500sShort": "500s",
-            "600sShort": "600s",
-            "700sShort": "700s",
-            "800sShort": "800s",
-            "900sShort": "900s",
-            etcShort: "ETC",
-        },
-        term: {
-            "3yearsShort": "In-3-Years",
-            "1yearShort": "In-1-Year",
-            thisSemesterShort: "This-Semester",
-        },
-        professors: {
-            over: " and ",
-            people: " others",
-        },
-        review: {
-            writingPlaceholder: "Please describe your evaluation of grades, load, etc.",
-            like: "Like ",
-        },
-        search: {
-            search: "Search",
-            cancel: "Cancel",
-            level: "Grade",
-            department: "Department",
-            term: "Term",
-            all: "All",
-            type: "Groups",
-            placeholder: "Enter keywords",
-            empty: "Please select at least one search filter",
-        },
-        semesters: {
-            spring: "Spring",
-            summer: "Summer",
-            fall: "Fall",
-            winter: "Winter",
-        },
-
-        days: {
-            monday: "Mon",
-            tuesday: "Tue",
-            wednesday: "Wed",
-            thursday: "Thu",
-            friday: "Fri",
-            saturday: "Sat",
-            sunday: "Sun",
-        },
-    },
-    header: {
-        dictionary: "Dictionary",
-        writeReviews: "Write Reviews",
-        timetable: "Timetable",
-    },
-    main: {
-        hisTimeTable: "<name>{{name}}</name><normal>'s Timetable</normal>",
-        schedule: {
-            title: "Due to {{content}}",
-            link: "Go to Student Portal",
-            contents: {
-                beginning: "Beginning",
-                end: "End",
-                courseRegistrationPeriodStart: "Start of Course Registration Period",
-                courseRegistrationPeriodEnd: "End of Course Registration Period",
-                courseAddDropPeriodEnd: "End of Course Add & Drop Period",
-                courseDropDeadline: "Deadline to Course Drop",
-                courseEvaluationDeadline: "Deadline to Course Evaluation",
-                gradePosting: "Grade Posting",
-            },
-        },
-        reviewSection: {
-            title: "How did you find the {{lectureName}} course?",
-            placeholder:
-                "How did you find the gradings and workload — fair, heavy, or just right?",
-
-            writeAnother: "Another one",
-        },
-        recentFeed: {
-            title: "<bold>Fresh</bold><space/><normal>Reviews</normal>",
-        },
-        likedMajorFeed: {
-            title: "<bold>Popular</bold><space/><normal>Reviews</normal>",
-        },
-        hallOfFameFeed: {
-            title: "<bold>Hall of</bold><space/><normal>Fame</normal>",
-        },
-        scheduleFeed: {
-            title: "School Schedules",
-        },
-    },
-    dictionary: {
-        courseHistory: "Course History",
-        notOffered: "Not offered",
-        courseCountInfo:
-            "Total<space/><bold>{{count}}</bold><space/>courses<space/>(<icon/>: offered this semester)",
-        sort: "Sort",
-        noResults: "No results",
-        sortOptions: {
-            code: "By Course Code",
-            popularity: "By Popularity",
-            studentCount: "By Enrollment",
-        },
-        review: "Course Reviews",
-        reviewLanguage: "Language",
-        reviewLanguageOptions: {
-            all: "All",
-            english: "English",
-        },
-    },
-    account: {
-        title: "My Account",
-        name: "Name",
-        email: "Email",
-        schoolId: "School ID",
-        course: "Course",
-        major: "Major",
-        interestedDepartments: "Interested Departments",
-        logout: "Logout",
-        sparcsssoCaption: "This information can be changed in <bold>SPARCS SSO</bold>",
-        kaistPortalCaption:
-            "This information is synced with KAIST Portal. If the information does not match, please contact <bold>otlplus@sparcs.org</bold>",
-    },
-    writeReviews: {
-        mySummary: {
-            title: "Taken Lectures",
-            written: "My Reviews",
-            likes: "likes",
-        },
-        write: {
-            title: "Write Reviews - {{lectureName}}",
-            related: "Related Reviews - {{lectureName}}",
-        },
-        hallOfFameFeed: {
-            title: "Hall of Fame - {{year}}<space/>{{semester}}",
-            total: "Total Reviews",
-        },
-        likedReviews: {
-            title: "My Liked Reviews",
-        },
-        tabs: {
-            write: "Write Reviews",
-            hallOfFameFeed: "Hall of Fame",
-            recentFeed: "Fresh Reviews",
-            liked: "My Liked Reviews",
-        },
-    },
-}
