--- app/i18n/variables/ko.ts
@@ -1,5 +1,0 @@
-import { type CommonI18nBaseType, commonI18nBase } from "./_base"
-
-export const commonI18nKo: CommonI18nBaseType = {
-  ...commonI18nBase,
-}
